# docs/swagger.json
{
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "DeFi 资产管理系统 API 文档",
        "title": "Rotki Demo API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "email": "support@example.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
...
        "/api/v1/rpc-nodes/{id}": {
            "put": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "rpc-nodes"
                ],
                "summary": "Update RPC node",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "RPC node ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Updated RPC node data",
                        "name": "node",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_rotki-demo_internal_models.RPCNode"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_rotki-demo_internal_models.RPCNode"
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "rpc-nodes"
                ],
                "summary": "Delete RPC node",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "RPC node ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    }
                }
            }
        },
        "/api/v1/rpc-nodes/{id}/check": {
            "post": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "rpc-nodes"
                ],
                "summary": "Check RPC node connection",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "RPC node ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
...
        "/rpc-nodes": {
            "get": {
                "description": "获取所有 RPC 节点，可按链 ID 过滤",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "rpc-nodes"
                ],
                "summary": "获取 RPC 节点列表",
                "parameters": [
                    {
                        "type": "string",
                        "description": "按链 ID 过滤",
                        "name": "chain_id",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/github_com_rotki-demo_internal_models.RPCNode"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "post": {
                "description": "创建一个新的 RPC 节点",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "rpc-nodes"
                ],
                "summary": "创建 RPC 节点",
                "parameters": [
                    {
                        "description": "RPC 节点数据",
                        "name": "node",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_rotki-demo_internal_models.RPCNode"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/github_com_rotki-demo_internal_models.RPCNode"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/rpc-nodes/{id}": {
            "get": {
                "description": "根据 ID 获取 RPC 节点",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "rpc-nodes"
                ],
                "summary": "获取 RPC 节点",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "RPC 节点 ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_rotki-demo_internal_models.RPCNode"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "github_com_rotki-demo_internal_models.Chain": {
            "type": "object",
            "properties": {
                "chain_type": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "is_active": {
                    "type": "boolean"
                },
                "logo_url": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "native_token_id": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
...
        "github_com_rotki-demo_internal_models.RPCNode": {
            "type": "object",
            "properties": {
                "chain": {
                    "description": "关系",
                    "allOf": [
                        {
                            "$ref": "#/definitions/github_com_rotki-demo_internal_models.Chain"
                        }
                    ]
                },
                "chain_id": {
                    "description": "eth、bsc、polygon 等",
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "is_connected": {
                    "description": "连接状态",
                    "type": "boolean"
                },
                "is_enabled": {
                    "description": "此节点是否处于活动状态",
                    "type": "boolean"
                },
                "last_checked": {
                    "description": "最后连接检查时间",
                    "type": "string"
                },
                "name": {
                    "description": "显示名称，如 \"0xRPC\"、\"PublicNode\"",
                    "type": "string"
                },
                "priority": {
                    "description": "优先级更高的节点优先",
                    "type": "integer"
                },
                "timeout": {
                    "description": "请求超时时间（秒）",
                    "type": "integer"
                },
                "updated_at": {
                    "type": "string"
                },
                "url": {
                    "description": "RPC 端点 URL",
                    "type": "string"
                },
                "weight": {
                    "description": "负载均衡权重（0-100）",
                    "type": "integer"
                }
            }
        },
...
    }
}